+++ src/class/ModelsCollection.ts
@@ -1,46 +1,44 @@
-import S3 = require('aws-sdk/clients/s3');
 import { IModel, IStorable, ISerilizable } from '../interface';
 import { Model, SingleSave } from './';
-import { IConfigModel } from '../interface/IObjects';
+import { ApiService, IApiModel } from '../service';
+import { Container } from 'typedi';
 
 export class ModelsCollection extends SingleSave implements IStorable, ISerilizable<IModel[], Model[]> {
 
   /** @type {Model[]} The list of model instances */
   private models: Model[];
-  /** @type {Model[]} The full path to the file */
-  private s3service: S3;
+  /** @type {Model[]} Remote API service */
+  private apiService: ApiService;
   /** @type {string} The pseudo path */
   public path: string;
   /** @type {string} The loaded instances */
   private static instances: ModelsCollection[] = [];
+  /** @type {string} The loaded instances */
+  private hashes: { [id: string]: string } = {};
 
   /**
    * Constructor
-   * @param {IConfigModel} config
+   * @param {string} project
    */
-  private constructor(public config: IConfigModel) {
+  private constructor(public project: string) {
     super();
-    this.s3service = new S3({
-      region: config.region,
-      accessKeyId: config.key,
-      secretAccessKey: config.secret
-    });
-    this.path = ModelsCollection.path(config);
+    this.apiService = Container.get(ApiService);
+    this.path = ModelsCollection.path(project);
   }
 
   /**
    * Returns a singleton for this config
-   * @param {IConfigModel} config
+   * @param {string} project
    */
-  public static async getInstance(config: IConfigModel) {
-    const path = ModelsCollection.path(config);
+  public static async getInstance(project: string) {
+    const path = ModelsCollection.path(project);
     // Try to find an existing collection
     const modelsCollection = ModelsCollection.instances.find((m) => m.path === path);
     if (modelsCollection) {
       return modelsCollection;
     }
     // Create and load a new collection
-    const collection = new ModelsCollection(config);
+    const collection = new ModelsCollection(project);
     await collection.load();
     // Keep the collection
     ModelsCollection.instances.push(collection);
@@ -50,37 +48,68 @@ export class ModelsCollection extends SingleSave implements IStorable, ISeriliza
 
   /** @inheritDoc */
   public async load(): Promise<void> {
-
-   const models = await this.s3service.getObject({
-      Bucket: this.config.bucket,
-      Key: this.config.path
+    const models = await this.apiService.get('model', {
+      _page: 0,
+      _limit: 100,
+      project: this.project
     })
-      .promise()
-      .then((data) => {
-        const content = (data.Body as Buffer).toString('utf8');
-        const models: IModel[] = JSON.parse(content);
-        this.didLoad(content);
-        return models;
-      })
-      .catch((error) => {
-        // First loading => no file => AccessDenied
-        if (error.code === 'NotFound' || error.code === 'AccessDenied') {
-          return [];
-        }
-        throw error;
+      .then(response => {
+        return (<IApiModel[]>response.data.items)
+          .map((m: IApiModel): IModel => ({
+            id: m._id,
+            name: m.name,
+            fields: m.fields,
+            accesses: m.accesses,
+          }));
       });
+
     this.fromObject(models);
+    this.updateHashes();
   }
   /** @inheritDoc */
   async save(): Promise<void> {
-    const data = JSON.stringify(this.toObject(), null, 2);
-    if (this.shouldSave(data)) {
-      await this.s3service.putObject({
-        Body: Buffer.from(data, 'utf8'),
-        Bucket: this.config.bucket,
-        Key: this.config.path
-      })
-        .promise();
+    
+    // Get models to create
+    const toCreate = this.models.filter(m => typeof this.hashes[m.id] === 'undefined');
+    
+    // Create models and update id
+    for(const model of toCreate) {
+      const response = await this.apiService.post('model', {
+        project: this.project,
+        name: model.name,
+        fields: model.fields,
+        accesses: model.accesses,
+      });
+      model.id = response.data._id;
+    }
+
+    // Get models to update
+    const toUpdate = this.models.filter(m => typeof this.hashes[m.id] === 'string' && this.hashes[m.id] !== m.hash());
+
+    // Update models
+    for(const model of toUpdate) {
+      await this.apiService.patch(`model/${model.id}`, {
+        name: model.name,
+        fields: model.fields,
+        accesses: model.accesses,
+      });
+    }
+
+    // Get models to delete
+    const toDelete = Object.keys(this.hashes).filter(id => !this.models.some(m => m.id === id));
+
+    // Delete models
+    for(const id of toDelete) {
+      await this.apiService.delete(`model/${id}`);
+    }
+    
+    this.updateHashes();
+  }
+  /** Update hashes from models */
+  private updateHashes() {
+    this.hashes = {};
+    for (const model of this.models) {
+      this.hashes[model.id] = model.hash();
     }
   }
   /**
@@ -114,7 +143,7 @@ export class ModelsCollection extends SingleSave implements IStorable, ISeriliza
    * Returns a pseudo path
    * @returns {string}
    */
-  private static path(config: IConfigModel): string {
-    return `s3:${config.bucket}:${config.path}`;
+  private static path(project: string): string {
+    return `project:${project}`;
   }
 }