+++ src/config/Remote.ts
@@ -1,0 +1,6 @@
+import { IConfigRemote } from '../interface';
+
+export const ConfigRemote: IConfigRemote = {
+  // uri: 'http://api.dev.hapify.io/v1',
+  uri: 'https://api.staging.hapify.io/v1',
+};