+++ src/config/index.ts
@@ -1,2 +1,2 @@
 
-export * from './Presets';
+export * from './Remote';