+++ src/service/Options.ts
@@ -1,6 +1,7 @@
 import { Service } from 'typedi';
 import { Command, CommanderStatic } from 'commander';
 import * as Path from 'path';
+import { GlobalConfigService } from './GlobalConfig';
 
 @Service()
 export class OptionsService {
@@ -11,7 +12,7 @@ export class OptionsService {
   private command: Command;
 
   /** Constructor */
-  constructor() {
+  constructor(private globalConfigService: GlobalConfigService) {
   }
 
   /**
@@ -28,10 +29,7 @@ export class OptionsService {
   setCommand(command: Command): void {
     this.command = command;
   }
-  /**
-   * Return the working directory computed with the --dir option
-   * @return {string}
-   */
+  /** @return {string} Return the working directory computed with the --dir option */
   dir(): string {
     if (this.program.dir) {
       if (Path.isAbsolute(this.program.dir)) {
@@ -40,6 +38,14 @@ export class OptionsService {
       return Path.resolve(process.cwd(), this.program.dir);
     }
     return process.cwd();
+  }
+  /** @return {string} Return the API Key to use (explicit or global) */
+  apiKey(): string {
+    const key = this.program.key || this.globalConfigService.getData().apiKey;
+    if (!key) {
+      throw new Error('Please define an API Key using command "hpf config" or the option "--key"');
+    }
+    return key;
   }
   /** @return {boolean} Denotes if the debug mode is enabled */
   debug(): boolean { return !!this.program.debug; }