+++ src/service/index.ts
@@ -1,6 +1,9 @@
 
 export * from './String';
+export * from './Api';
 export * from './Channels';
+export * from './Generator';
+export * from './GlobalConfig';
 export * from './Presets';
 export * from './Options';
 export * from './Logger';
@@ -9,17 +12,4 @@ export * from './HttpServer';
 export * from './Validator';
 export * from './WebSocketServer';
 
-export * from './generators/JsGenerator';
-export * from './generators/HpfGenerator';
-export * from './generators/Generator';
-
-export * from './websocket-handlers/GetModelsHandler';
-export * from './websocket-handlers/SetModelsHandler';
-export * from './websocket-handlers/GetChannelsHandler';
-export * from './websocket-handlers/SetChannelsHandler';
-export * from './websocket-handlers/GetPresetsHandler';
-export * from './websocket-handlers/PathPreviewHandler';
-export * from './websocket-handlers/TemplatePreviewHandler';
-export * from './websocket-handlers/ValidateModelHandler';
-export * from './websocket-handlers/GenerateTemplateHandler';
-export * from './websocket-handlers/GenerateChannelHandler';
+export * from './websocket-handlers';