+++ src/service/websocket-handlers/SetModelsHandler.ts
@@ -1,6 +1,6 @@
 import { Service } from 'typedi';
 import { WebSocketMessages, IWebSocketHandler, IWebSocketMessage } from '../../interface';
-import { ChannelsService } from '../';
+import { ChannelsService } from '../Channels';
 import * as Joi from 'joi';
 import { Access } from '../../interface/IObjects';
 
@@ -51,7 +51,7 @@ export class SetModelsHandlerService implements IWebSocketHandler {
         search: Joi.string().valid(accesses).required(),
         count: Joi.string().valid(accesses).required(),
       })
-    })).min(1);
+    })).min(0);
   }
 
   /** @inheritDoc */