+++ CHANGELOG.md
@@ -6,8 +6,12 @@
 - Handle scenarios.
 
 ### Changed
+- Use remote models storage.
+- Use remote scenarios storage.
+- Use remote generator.
 
 ### Removed
+- Remove local generator.
 
 ## [0.3.0] - 2018-11-15